+++ app/frontend/src/javascript/components/fab-modal.tsx
@@ -1,8 +1,8 @@
 /**
- * This component is a modal dialog that can wraps the application style
+ * This component is a template for a modal dialog that wraps the application style
  */
 
-import React from 'react';
+import React, { ReactNode } from 'react';
 import Modal from 'react-modal';
 import { useTranslation } from 'react-i18next';
 import { Loader } from './loader';
@@ -13,14 +13,22 @@ Modal.setAppElement('body');
 interface FabModalProps {
   title: string,
   isOpen: boolean,
-  toggleModal: () => void
+  toggleModal: () => void,
+  confirmButton?: ReactNode
 }
 
 const blackLogoFile = CustomAssetAPI.get('logo-black-file');
 
-export const FabModal: React.FC<FabModalProps> = ({ title, isOpen, toggleModal, children }) => {
+export const FabModal: React.FC<FabModalProps> = ({ title, isOpen, toggleModal, children, confirmButton }) => {
   const { t } = useTranslation('shared');
   const blackLogo = blackLogoFile.read();
+
+  /**
+   * Check if the confirm button should be present
+   */
+  const hasConfirmButton = (): boolean => {
+    return confirmButton !== undefined;
+  }
 
   return (
     <Modal isOpen={isOpen}
@@ -40,7 +48,8 @@ export const FabModal: React.FC<FabModalProps> = ({ title, isOpen, toggleModal, 
       </div>
       <div className="fab-modal-footer">
         <Loader>
-          <button className="close-modal-btn" onClick={toggleModal}>{t('app.shared.buttons.close')}</button>
+          <button className="modal-btn--close" onClick={toggleModal}>{t('app.shared.buttons.close')}</button>
+          {hasConfirmButton() && <span className="modal-btn--confirm">{confirmButton}</span>}
         </Loader>
       </div>
     </Modal>